requirements.txt
python-dotenv
supabase
psycopg2-binary